Add initial example of context API and logger handler.
