+++ timber/__init__.py
@@ -1,7 +1,6 @@
 # coding: utf-8
 from __future__ import print_function, unicode_literals
 from logging import Handler, Formatter, NOTSET
-from functools import wraps
 import base64
 import json
 import requests
@@ -9,29 +8,18 @@ import pdb
 from timber.schema import validate
 from datetime import datetime
 
+from timber.helpers import _debug
+from timber.context import ContextLogger
+from timber.constants import TIMBER_CONTEXT_KEY
 
-def _debug(*data):
-    print(json.dumps(data, indent=2, sort_keys=True))
 
-
-def _collapse_contexts(fn):
-    @wraps(fn)
-    def wrapped(self, *args, **kwargs):
-        x = {}
-        for name, e in self.extras:
-            x.setdefault(name, {}).update(e)
-        if 'extra' in kwargs:
-            x.setdefault('extra', {}).update(kwargs['extra'])
-        if x:
-            kwargs.setdefault('extra', {})['__timber_context'] = x
-        return fn(self, *args, **kwargs)
-    return wrapped
+def _context_from_record(record):
+    return record.__dict__.get(TIMBER_CONTEXT_KEY, None)
 
 
 def _create_payload(record):
     payload = {}
     record_dict = record.__dict__
-    pdb.set_trace()
 
     payload['dt'] = datetime.fromtimestamp(record_dict['created']).isoformat()
     payload['level'] = record_dict['levelname'].lower()
@@ -44,7 +32,7 @@ def _create_payload(record):
     runtime['file'] = record_dict['filename']
     runtime['line'] = record_dict['lineno']
     # Custom context
-    custom_context = record_dict.get('__timber_context', None)
+    custom_context = _context_from_record(record)
     if custom_context:
         context['custom'] = custom_context
     return payload
@@ -80,28 +68,3 @@ class TimberHandler(Handler):
             e.response = response
             raise e
         _debug(response.status_code, response.content)
-
-
-class ContextLogger(object):
-    def __init__(self, logger, api_key):
-        self.logger = logger
-        self.handler = TimberHandler(api_key)
-        self.logger.addHandler(self.handler)
-        self.extras = []
-
-    def context(self, name, ctx):
-        self.extras.append((name, ctx))
-        return self
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, type, value, traceback):
-        if type is not None:
-            return False
-        self.extras.pop()
-        return self
-
-    @_collapse_contexts
-    def critical(self, *args, **kwargs):
-        return self.logger.critical(*args, **kwargs)
