+++ timber/constants.py
@@ -1,0 +1,5 @@
+# coding: utf-8
+from __future__ import print_function, unicode_literals
+
+
+TIMBER_CONTEXT_KEY = '__timber_context'
